Created organizationData routes

+++ backend/routes/organizationData.js
@@ -1,5 +1,3 @@
-// TODO: ADD BETTER ERROR HANDLING WITH STATUS CODES
-
 // import express module
 const express = require("express");
 const router = express.Router();
@@ -20,42 +18,56 @@ router.get("/",(req,res,next) => {
 
 // POST new organizations
 router.post("/",(req,res,next)=>{
-    orgdata.create(
-        req.body,
-        (error, data) => {
-            if (error){
-                return next(error);
-            } else {
-                res.json(data);
+    if (req.body._id) {
+        orgdata.create(
+            req.body,
+            (error) => {
+                if (error){
+                    return next(error);
+                } else {
+                    res.send(req.body._id + " has been added successfully.");
+                }
             }
-        }
-    );
+        );
+    } else {
+        res.status(400).send("ERROR: No organization ID provided.");
+    }
+    
 });
 
 // PUT or Update organizations
 router.put("/:id",(req, res,next) => {
-    orgdata.findOneAndUpdate(
-        {_id: req.params.id}, req.body,(error, data) => {
-            if (error) {
-                return next(error);
-            } else {
-                res.json(data);
+    if (req.params.id){
+        orgdata.findOneAndUpdate(
+            {_id: req.params.id}, req.body,(error, data) => {
+                if (error) {
+                    return next(error);
+                } else {
+                    res.send(req.params.id + " has been updated successfully.");
+                }
             }
-        }
-    );
+        );
+    } else {
+        res.status(400).send("ERROR: No organization ID provided.")
+    }
 });
 
 // DELETE organization
 router.delete("/:id",(req,res,next)=>{
-    orgdata.findByIdAndDelete(
-        {_id:req.params.id}, req.body,(error,data) => {
-            if (error) {
-                return next(error);
-            } else {
-                res.json(data);
+    if (req.params.id){
+        orgdata.findByIdAndDelete(
+            {_id:req.params.id}, req.body,(error,data) => {
+                if (error) {
+                    return next(error);
+                } else {
+                    res.send(req.params.id + " has been deleted successfully.")
+                }
             }
-        }
-    );
+        );
+    } else{
+        res.status(400).send("ERROR: No organization ID provided.")
+    }
+    
 });
 
 module.exports = router;
